requirement.txt
streamlit
streamlit-chat
langchain
langchain-google-genai
google-generativeai
faiss-cpu
python-dotenv
